[HDRP] Fix visible light clamping in path tracer with low exposure setting
This fixes an issue in the HDRP path tracer that is mainly visible with low exposure settings (-5 for instance) and a neutral tone mapper, where the affected region of a light seems to be clamped to a spherical volume, even though the range and intensity of the light should lit the regions outside of that sphere. This was due to an epsilon threshold value on the light intensity that did not take into account the exposure value. The epsilon is also lowered for volume scattering as the effect was still slightly visible, even with exposure compensation.

Here are two screenshots showing the effect on volumetric fog and surface:
![BugFogRadius_20231117_1036760_SurfaceBug](https://media.github.cds.internal.unity3d.com/user/4198/files/a1b30420-2a5d-4e2c-9cdd-4221ea928361)
![BugFogRadius_20231117_1016612_VolumeBug](https://media.github.cds.internal.unity3d.com/user/4198/files/d4c7bc37-5509-4734-b251-2a8c81365b1a)

And here are screenshots with the fixed behaviour:
![BugFogRadius_20231117_1037666_SurfaceFix](https://media.github.cds.internal.unity3d.com/user/4198/files/a9d5a802-b9cb-4fa9-b280-a70cb8a3cc87)
![BugFogRadius_20231117_1015377_VolumeFix](https://media.github.cds.internal.unity3d.com/user/4198/files/061998b9-c5ee-4913-abe1-2ac8f7ee2b89)