Gpu lightmap batch breaker
This PR adds an option to disable lightmap texture arrays when GPU-driven rendering is enabled. Previously texture arrays were always enabled. This remains the default.

![image](https://media.github.cds.internal.unity3d.com/user/7119/files/5fcd814c-5823-4790-9ce0-4576ece3a30d)

This is a tradeoff. When the option is disabled, lightmaps are bound as texture arrays. This reduces the number of draw calls, but it may perform worse on some GPUs. When the option is enabled, lightmaps are bound as individual textures. This breaks the batch whenever new lightmaps need to be bound, which can result in an increased number of draw calls, but it consumes less memory. It also performs better on some lower-end GPUs where dependent reads from texture arrays can incur large performance penalties.

![image](https://media.github.cds.internal.unity3d.com/user/2726/files/b43f06f0-f951-493b-b201-932213f0f7e1)

https://jira.unity3d.com/browse/POI-553