2022.3/2d/staging/tilemap/uum55308
*[See [this document](https://docs.google.com/document/d/1C4Vc7dOKgoZ2mNqMdugHDg-w6iN7iUYdlBQMGDmVS3E/edit) for details on what can land in 2023.3.]*

This PR fixes an issue in the TilemapRenderer where it picks up and uses texture shader properties from the previous render node and uses it, when the TilemapRenderer does not set the shader properties. Instead, a default texture property will be used if the shader property is not set in the TilemapRenderer material.