[APV] Bake probe volumes using light transport API
https://jira.unity3d.com/browse/HDRP-3119

Updated APV baking to use the new light transport API, which gives more control on the baking process.
This simplified A LOT the baking process as we are now less dependant of the lightmapper callbacks and we can manage stuff more clearly.

The main advantage is that we can control the sample count used during the bake, and therefore have an adjustment volume to locally increase the sample count in interesting areas

![image](https://media.github.cds.internal.unity3d.com/user/2154/files/3b00fe14-715f-4969-a33b-0f1c323a1abd)

This also allows to trigger bakes without launching a 'real bake', which is used for two things now:
- Bake only probe volumes. This doesn't bake the lightmaps, so it allows faster iteration when working on APV
![image](https://media.github.cds.internal.unity3d.com/user/2154/files/2a341b88-86e6-4ca1-a8d7-d9c4cae402a6)
- Bake reflection normalization. When baking a reflection probe via it's inspector, we can trigger a bake of 1 probe to update the normalization coefficients
![image](https://media.github.cds.internal.unity3d.com/user/2154/files/4597bff0-5420-4a6f-9e13-d32edc86a688)
- TODO: We can now more easily recompute a subset of probes, like inside an adjustment volume. This is not straightforward as we need to merge partial bake results so it will be in a separate PR. Also we need to figure out with david the best UX for that

Note: the c++ changes were authored with kasper (see commit list) so i didn't add him as reviewer
Note 2: i had to refactor a lot of the sky occlusin baking code so it's diffuclt to review, sorry about that