[Backport 2022.3][Jira # UUM-52109] Fixing half res transparent hardware DRS off by 1
Fixing issues with hardware DRS on console (manifestation is usually bright qnan pixels on the right of the screen). 
Lowres transparent (aka half res transparent) can be off by one pixel with the hardware resolution dynamic scale. The reason is a small mismatch between RTHandle and the runtime. Unfortunately fixing this problem might require a heavier rewrite to ensure the RTHandle scale works well. As a fix we can compute the error of this scale and plumb it in as a custom DRS scaler (only for the HW path) and fix the issue. Issue only present on xbox as qnans in the border (triggered by only certain resolutions using hardware DRS with transparent lowres).