[Graphics Test Framework] Force ordinal sorting when generating the Shader Variant lists
With https://github.cds.internal.unity3d.com/unity/unity/pull/38912 we are automating the generation of the Shader Variant lists. It's currently done manually by the developers using the graphics test framework.
However the sorting algorithm used to generate the list is not set explicitly, it's left to the framework (unity/.net) and local implementation of the SortedSet. 

This PR is making the SortedSet use `StringComparer.Ordinal` explicitly, so that the manual workflow through the GTF has the same outcome as the automated workflow from https://github.cds.internal.unity3d.com/unity/unity/pull/38912.

This PR also updates the existing .shadervariantlist files as a result. A lot of duplicates were also removed automatically.