UUM-56914 #comment Fixed BiRP shader graph target displaying specular…
[Jira Bug](https://jira.unity3d.com/browse/UUM-56618)
[Port](https://jira.unity3d.com/browse/UUM-56914)

The SG BiRP target was incorrectly showing a specular block when material override was enabled. Specular workflow isn't supported (and can't be picked) so this was incorrect behavior. A comment was added to explain why material override was displaying the value even when the workflow mode wasn't set and the offending code was removed with a comment.