[Jira # UUM-57601] Read / Write hack fix for async compute issues on HDRP console
Async compute Presents issues when depending on the first texture read of depth. In some platforms (vulkan/dx12 and some consoles) the first read of depth is actually a write because of decompression / layout / barriers.
Async compute is inserting the fence before this operation because RenderGraph sees everything as a simple Read / Write.
Temporarily hacking this so the first read of depth is tagged as a write, so the correct fence is inserted.