Fix UniversalRendererResources settings getting incorrectly stripped from player builds
This PR fixes https://jira.unity3d.com/browse/UUM-57954.

During player build, in some circumstances, potentially depending on whether viewport is visible, or if editor is run in headless mode, the `ScriptableRenderer` objects returned by `UniversalRenderPipelineAsset.renderers` may be null. If this happens during the build process, the `RendererStripper` would strip the settings from the player, resulting in missing shaders at runtime (such as CopyDepth shader).

This is fixed by doing the checks using `ScriptableRendererData` objects instead, which are part of the asset's serialized state and not constructed at runtime.

Also fix a warning caused by an abstract SettingsStripper class, which shouldn't be a warning.