Fix prevProjMatrix Jitter in HDRP
This variable is not intended to be jittered but it was sourcing its data from a potentially jittered value. This was causing STP's static geometry reprojection results to be inaccurate in some projects which significantly degraded visual quality.

This issue was only happening in cases where the AfterPostProcess frame setting was enabled, so it didn't show up in projects such as HDRP Sample Scene, but it did show up in Spaceship Demo.