[UIT-2237]: Reverting the new Contextual Menu feature.
This PR reverts the new contextual menu back to native contextual menu. Our approach was to take the original PR, dissect it and revert all the necessary code to make this happen. 

It is important to highlight that some areas were affected by this changes. Specifically the Piercing Menu, done by the Scene tooling team and the Menus in the render pipeline stream. With this being said, this PR **will** contain code changes coming from the Scene Tooling and Render Pipeline team respectively. We will ensure to work together to get their affected test(s) and feature passing for 23.3.