[URP]Implementing uUniversalRenderPipelinePostProcessDataUsage Analytic.
Post Process Data is an asset referenced into the UniversalRenderer and the Renderer2D for URP.

![image](https://media.github.cds.internal.unity3d.com/user/3279/files/4c29f342-ed35-48a6-af9c-e47d18cdcb5e)


We do not know what are the usages of this asset, and how users are modifying it.

We want to know basically
1- Are users modifying the shaders 1 single time for the whole project?
2- Are users modifying the shaders per quality level?

With this information we will be able to make a decision if this asset is migrated to GraphicsSettings or we need to think of an alternative. 

Event sheet for Data team: https://docs.google.com/spreadsheets/d/1YqIZkUi0Y_pd5qpC97i1S-dF6CaL7s1kndi-uHlHs1E/edit#gid=1274740786