[URP][2023.3] Fixing an issue where SSAO feature can not be disabled in runtime (UUM-56165)
Fixes UUM-56165.
The issue is that when a project only contains renderers with the SSAO feature, the OFF variant is taken out to improve build times. This however presents an issue when the feature is disabled in runtime as the keyword is always enabled.

The fix here is to use one of the unused components of `_AmbientOcclusionParam` to specify whether the feature is enabled or not and use that with the calculated SSAO value.