[VFX] Particle strips missing functionality
This PR implements missing functionality for particle strips. In particular, it adds particle count readback, which enables sleep state, and per-particle frustum culling.

- Particle count readback:
Strips don't use dead lists to spawn new particles, so we can't reuse that counter, like we do with particle systems. Instead, we are adding some space in memory (1 uint per instance) to store the particle count, at the beginning of the strip data buffer.
In the UpdateStrips compute we were already computing the particle count for each strip, so we added an InterlockedAdd to sum the counts from all the strips of each instance, storing it in the new reserved space.
We also need to reset the counters to 0 every frame. We do this in the particle update, which was already writing to the StripData buffer. This way, we avoid setting the data from CPU, which is what we do for other counters.
There is one difference compared to particle systems readback. For strips, we are returning the actual particle count, while for regular particle systems we return the number of available particles. The readback system takes care of it, substracting from capacity only for particle systems.

- Sleep state:
One important advantage of adding the missing particle count readback is that we are now able to detect when strip systems are empty (0 particles and not expected to spawn) and we can set it to sleep, reusing the implementation that we already have for particle systems.

- Frustum culling:
Most of the implementation for frustum culling was already possible to reuse after implementing indirect draw.
In this PR we remove the code that was disabling this option, and modify the bounding sphere computation for strip particles, including the next particle that the strip will be connected to.