Graphics/SRP/RPF - RG Opt 3 - Disable Render Graph Validation Checks on demand
Render Graph does a lot of validation checks to ensure resources are not corrupted and correctly setup. These checks are costly. Disabling them provides 0.1-0.2ms gains on main CPU thread on PS4 and Switch with High Quality mode tests. This PR implements a way to enable/disable the validation layers from the UI in the editor for both URP and HDRP. Validations will always be disabled in release mode.