[UUM-53816][SRPCore]Fixing NullReferenceException when trying to display VolumeComponents overrides.
The VolumeManager initializes the volumes for a pipeline when it is initialized.
In the user project, even if the scene view was open, the UniversalRenderPipeline was not created as there was a Built-in debug mode active. When adding an Override on the Volume Profile, the list was null and it has never been initialized. 

Adding a fetch of the Volume Component types even if the Volume Manager has not been initilazed.

Fixing some usages of `GetVolumeComponentsForDisplay`, as it should be RenderPipelineAsset type, and not RenderPipeline type.