[VFX] Add instancing support to ShaderGraph keywords feature
JIRA: https://jira.unity3d.com/browse/VFXG-414
Epic: https://jira.unity3d.com/browse/VFXG-408
POI: https://jira.unity3d.com/browse/POI-816

This PR is adding instancing support to the recently landed ShaderGraph keywords feature:
https://github.cds.internal.unity3d.com/unity/unity/pull/38694

Without instancing support this feature will not scale properly and won't be production ready.

The implementation of this feature is transparent to the user, the only noticeable difference is that the VFX asset will not display the "Instancing disabled" warning, and better performance when using it with several instances.

To support this, we are creating one SharedMaterialData for each combination of keywords that is used, for each material. Then, when adding the rendering node in VFXRenderer, we check if this instance needs a SharedMaterialData different than the one used in the material, and replace it in the materialInfo.

Overhead for particle systems not using keywords should be minimal.