[HDRP][Denoising] Remove broken denoising caching mechanism to avoid Editor freeze in certain cases
https://jira.unity3d.com/browse/UUM-57582

Our dirtiness check for the path tracer had a special case to avoid doing a full path tracing reset when you change denoising parameters and you have enough data to run denoising again without any new data. But this is very fragile as we increase the complexity of the denoising solution (for example with the volumetrics) so at this point it seems that it's better to remove this code and always do a reset.