[SRP] Automate ShaderVariantList update
https://jira.unity3d.com/browse/ANT-962

The shader variant stripper allows huge gains on the shader compilation of SRP test projects, allowing us to run standalone tests much faster, and sometimes to have them in the QV. 
- The feature was introduced here: https://github.cds.internal.unity3d.com/unity/com.unity.testframework.graphics/pull/146
- Design doc: https://docs.google.com/document/d/1xnkDTeMNZ_MFIVBWzfGr6nfWmKUj07rjP31vWZKITfU/edit#heading=h.v4zit0fymxn7
TL;DR: The feature is working by providing the explicit list of shader variants that are actually used by the graphics tests (as compiling the whole shader variants space is too long).

One pain point currently is for the developers to generate said lists. It's happening manually, by opening a unity project with the graphics test framework and using the inspector. It's very tedious to do it for each test project, and for each platform.
This PR and initiative are about automating this update.

A new C# solution was introduced, with 2 projects:
- ShaderVariantListUpdater, containing the actual features.
- ShaderVariantListUpdater.Tests, containing unit tests.

A new Yamato job was also introduced, as well as boilerplate for code ownership, config, etc.
The yamato job takes a `Pipeline ID` and `Update mode` as input from the developers that kicks off the job.

### How does the updater work?

**Inputs:**
- `Pipeline ID`, pointing to a pipeline holding the jobs running the test projects we want to update the shadervariant lists from ([example](https://unity-ci.cds.internal.unity3d.com/job/32039877/dependency-graph))
- `Update mode`: either `update` or `aggregate`, by default update. If update is specified it will override the existing shadervariantlist file, if aggregate is specified it will append to it.

**Workflow:**
1. Instantiate of a Services provider, that we'll use for Dependency Injection (DI) throughout the lifecycle of the process (AppSettings, Logger)
2. Trigger the Updater service
3. Fetch the Pipeline object and its dependencies with the Yamato REST API
4. Download and unzip the relevant artifact (in this case, test execution logs) in the temp local folder
5. Foreach job (each dependency from 3):
    - Delete the existing shadervariantlist file on the first pass (if `UpdateMode == Update`) 
    - Parse the existing shadervariant list file (after the first pass) into a SortedSet (to get unique values)
    - Pass the log to the SVLGeneration Service, which will use regexes to match all relevant shader variants and add them to the set.
    - Dump the SortedSet content to the shadervariantlist file

#### How does the yamato job work?
Using the updater is possible locally, but there's a yamato job that runs to by default match the HDRP jobs (can be changed to e.g. URP). The job simply triggers the program and commits/pushes the shadervariantlist file updates to the dev branch.

**Caveats**
- A prerequisite of this job is to have a "valid" standalone pipeline run to input it ([example](https://unity-ci.cds.internal.unity3d.com/job/32039877/dependency-graph))
- By valid we mean with a very low number of jobs failing in that pipeline. For shadervariantlist updates to be relevant we need as many platforms and APIs as possible.