[UUM-52949][URP][Bugfix] MSAA alpha to coverage issue
This PR fixes the bug case: UUM-52949.
The issue was caused by the previous frame not being cleared before drawing opaque objects with alpha clipping enabled. This resulted in an incorrect pixel blending for alpha to coverage. The error is most noticeable when using a transparent object with additive blending since the pixel becomes really bright in the blending area.

We can't make a complete solution for the blending without breaking changes but we can avoid making the issue standout or use last frames pixels by clearing the screen in MSAA is used at the start of the rendering of a frame.