[APV] Fix crash when baking sky occlusion
When baking a very large scene, sky occlusion issues a dispatch too large result in an error in the console and bake fails.
Can be easily reproed with a big terrain, eg. this one would fail on trunk
![image](https://media.github.cds.internal.unity3d.com/user/2154/files/9a086574-b2ba-4243-a72d-ad5e78f64086)

Note: a recent change from arthur (#40707) hides the option to cull backfaces on sky occlusion, but doesn't force the value to false, so it keeps the value it had before the option was hidden. This is unwanted and so now the value is always forced to false (explanation of why can be found in the original PR)