[HDRP] Fix usage of max z buffer for volumetric clouds
reported by the demo team here https://unity.slack.com/archives/C022VGK09HT/p1702945817869849


The max z buffer was used in distant clouds mode and is not useful anymore as we use directly the regular depth buffer (it's more precise obviously and faster than generating it when volumetric fog is disabled)
But the code sampling the max z buffer was left making blocky artifacts in some cases
Can be reproduced by putting a cube intersecting volumetric clouds and enabling volumetric fog. Blocky artefacts will appear

| before | after |
|--|--|
| <img width="509" alt="image" src="https://media.github.cds.internal.unity3d.com/user/2154/files/6dde9f7a-3f11-4fd4-9cca-bcfa73413ea1"> | <img width="509" alt="image" src="https://media.github.cds.internal.unity3d.com/user/2154/files/69d7838a-b675-428e-83ab-20cc8dab3570"> |

Also fix an issue with the integration noise when camera is in the clouds

| before | after |
|--|--|
| ![image](https://media.github.cds.internal.unity3d.com/user/2154/files/dd5e1bf5-f357-4aff-8f69-0472294ce352) | ![image](https://media.github.cds.internal.unity3d.com/user/2154/files/aedae82e-87d9-4851-a269-c34e6e23fbc3) |

Finally fixes an issue on clouds when fog is enabled and camera is a bit high. They would end up further than the far plane and receive too much fog
| before | after |
|--|--|
| ![image](https://media.github.cds.internal.unity3d.com/user/2154/files/bc571253-91d9-4604-89be-a3c10734237d) | ![image](https://media.github.cds.internal.unity3d.com/user/2154/files/2da7e9c2-f083-4cba-87c5-fb86cd819c61) |