Fix Swapped Luma Weights in STP & Other Small Typos
This PR addresses some typos that were present in the previous version of the shader
code in Stp.hlsl.

Specifically it fixes:

- Swapped constants for color channels in luma calculation
- Compilation issues in HLSL 2021 mode due to vector ternary usage
- Incorrect array index used in dilation constant code (unused in Unity's implementation)

The luma weight changes will cause a subtle change in visual output, but the overall impact
isn't expected to be noticeable. The other changes don't affect runtime functionality.