Fix Inconsistent STP Naming in SRPs
The code and several places in the UI still refer to STP by its name before it was changed for the Unite demo. This change corrects the code and UI to follow the naming convention that was used when the feature was announced to the public.

Before Demo: "Scalable Temporal Post-Processing"
After Demo: "Spatial-Temporal Post-Processing"