Fix missing unregistration of deleted diffusion profiles from Default Volume Profile
This PR fixes https://jira.unity3d.com/browse/UUM-55807.

The code in `VolumeUtils.TryAddDiffusionProfiles` was already removing null entries from the Default Volume Profile in an asset post-processing callback. This PR removes the early-exits that prevented this function from being called if nothing is being added. So now we always call it and cleanup null entries immediately as they appear, which fixes the bug.