[HDRP] Fix Compilation Error in SixWayLit VFX
The GetSixWayDiffuseContributions function expects an input array defined with the "real" type. When building for platforms that support reduced precision, this type is actually a min16float. Normally this would just cause a warning, but in this case, the array type cannot be automatically type casted so it results in an error which prevents several VFX Render Quad shaders from compiling successfully.

This change modifies the calling code to use the "real" type in order to guarantee a matching type.