[Graphics][urp][bugfix] Early return on renderer features requiring color for offscreen depth cameras.
Fix the remaining Renderer Features for offscreen depth cameras that don't have the color buffer. This will allow the renderer to not throw an exception when the camera is outputting to a depth-only render texture. The remaining Renderer Features were FullScreenPassRendererFeature and the RenderObject.

I made a previous assumption for FullScreenPassRendererFeature that it could run without reading or writing to color which is incorrect. This PR also fixes my error.

Earlier PR: https://github.cds.internal.unity3d.com/unity/unity/pull/27096
Previous Jira: UUM-14679