Fix incorrect volume stack dispose (UUM-54394, UUM-35147)
Fix for https://jira.unity3d.com/browse/UUM-54394

Incidentally this seems to be the same issue that was originally reported and fixed with this bug https://jira.unity3d.com/browse/UUM-35147, but it was never backported to 21.3. So this PR is effectively the missing backport of https://github.cds.internal.unity3d.com/unity/unity/pull/28676 into 21.3.