[VFX] Fix cast shadows property in Mesh Output blocks not working properly on URP
In URP, Mesh Output contexts (VFXStaticMeshOutput) still cast shadows even when their "cast shadows" property is unchecked.

![image](https://media.github.cds.internal.unity3d.com/user/2805/files/0bb44af3-fa28-46db-b52e-d6fc96f71481)

The "cast shadows" property works by disabling the ShadowCaster pass, if it exists in the shader. However, in URP, this pass gets enabled/disabled **during material validation**, based on some conditions. For instance, if _CastShadows property exists, it will use that value. If it does not, but it is a ShaderGraph, shadows will be enabled.

```            // Cast Shadows
            bool castShadows = true;
            if (material.HasProperty(Property.CastShadows))
            {
                castShadows = (material.GetFloat(Property.CastShadows) != 0.0f);
            }
            else
            {
                if (isShaderGraph)
                {
                    // Lit.shadergraph or Unlit.shadergraph, but no material control defined
                    // enable the pass in the material, so shader can decide...
                    castShadows = true;
                }
                else
                {
                    // Lit.shader or Unlit.shader -- set based on transparency
                    castShadows = Rendering.Universal.ShaderGUI.LitGUI.IsOpaque(material);
                }
            }
            material.SetShaderPassEnabled("ShadowCaster", castShadows);
```


This is one of the few cases where VFX is using shaders that are not owned by it, it is not possible to modify directly the shader to control shadows. The original approach, which is used in HDRP as well, was disabling the ShadowCaster pass after the initial setup. But materials can be validated later as well, when they are created (VFX batches hold a copy of materials). This is currently reenabling the ShadowCaster pass outside VFX graph control.

The only solution that we could find is to add VFX code inside the material validation/setup, to ensure that the passes remain as they should.

The original bug that this PR is fixing is related to shadows, but we have exactly the same issue with motion vectors, that should always be disabled for these type of outputs. This PR also addresses that issue, ensuring that the MotionVector pass gets disabled when requested.