Fix GPU Resident Drawer error on unsupported build targets
This is a follow-up for #42620 to avoid an error when using the editor with an active build target that does not support the GPU Resident Drawer (such as GLES3 on Android).  This error was breaking some builders, such as `Build_Universal_BatchRendererGroup_Android_OpenGLES3_Standalone_il2cpp_Linear`.  The fix checks for support earlier, emitting an editor-only warning instead.