Fixed lens flare occlusion format support test.
Fixed the format test for Lens Flare occlusion texture (required format is not supported bellow DX11 caps).
This was originaly done in trunk in this PR https://github.cds.internal.unity3d.com/unity/unity/pull/27324 but we were not able to backport it so we're fixing this particular issue directly on 22.3.