[Water] Fix water in SSR reflections
Color pyramid was generated before rendering water lighting
But as water absorption on opaque is rendered during opaque atmospheric scattering pass, the color buffer contains water absorption, making black spots where there should be water.
To fix that, we render water lighting before color pyramid used for transparent SSR, so color pyramid is nice. This means however that we have to remove the color pyramid offset option when rendering underwater as it's not generated yet now
![image](https://media.github.cds.internal.unity3d.com/user/2154/files/b729bc96-64ae-4d2d-838e-6d48c656cf6e)

Note that as water is still not in the depth pyramid, the SSR depth is not correct and when camera is moving, water reflecting on objects is not very stable, but we don't really care i think

A nice bonus advantage of this change is that now the sample color node in SG used for waterline doesn't need distortion to be enabled anymore 🙂 

| trunk | this |
| -- | -- |
| ![image](https://media.github.cds.internal.unity3d.com/user/2154/files/ac0257c5-1088-45d9-b9f7-aeedcfe1ef10) | ![image](https://media.github.cds.internal.unity3d.com/user/2154/files/da750894-ea24-46bf-bf8b-42719f5e0245) |
| ![image](https://media.github.cds.internal.unity3d.com/user/2154/files/6a981b77-ef15-40d4-a54e-a1218e876b83) | ![image](https://media.github.cds.internal.unity3d.com/user/2154/files/1c8a2edc-a243-44a8-a494-bade812a12be) |