[HDRP] Fix to avoid crashes on some GPUs
On some platforms/GPUs, ternary operators in shader code might be rearranged and cause both sides to be evaluated before the condition is tested.
For some shader locations this can cause out-of-bound memory accesses and GPU crashes.

This PR fixes one such case (that happened when running the Spaceship demo on console).