Fix for using URP render scale with GPU occlusion
This PR is a fix for incorrect GPU occlusion culling (POI: https://jira.unity3d.com/browse/POI-926) when using render scale in URP.  This was due to using pre-scaled width and height when building occluders from the depth buffer, the fix is to use scaled dimensions.

This issue only affects URP, the HDRP integration was already using scaled dimensions.

GPU occlusion culling is a new feature in 2023.3 alpha, so backports are not required.

Before this fix at 70% render scale (occlusion culling is not consistent with the scene):

![image](https://media.github.cds.internal.unity3d.com/user/4787/files/40c2a2d0-d624-4df4-aed6-e315dcb9f345)

After the fix at 70% render scale (occlusion culling is correct):

![image](https://media.github.cds.internal.unity3d.com/user/4787/files/3c5e8bf7-66c2-46e9-8a11-616729835bba)