[2023.3][URP] Fixing an issue with flickering screen when using Rendering Layers in deferred (UUM-61041)
Fixes UUM-61041.
This issue reproduces when having both the scene and game windows open using a deferred renderer with rendering layers. The issue is because the Rendering Layers texture wasn't updated like the GBuffer 0 to 3. 

This PR changes it so RenderingLayers texture is fetched with RenderPass instead of sampling a texture.