[HDRP] Improvement to atmospheric scattering for space games
The goal of the PR is to adress feedback on the forum where user have difficulties getting nice results when making a game with a camera in space
https://forum.unity.com/threads/physically-based-sky.765962/page-6#post-9572743


Changing the default ShaderConfig for PrecomputedAtmosphericAttenuation to true.
This means sun light absorption from the atmosphere is computed on the CPU at the camera position, instead of per pixel during lighting. This is less correct but more performant. (saves a bit moremore than 0.5 ms on the fullscreen deferred lighting)
| Precomputed off (previous) | Precomputed on (new default) |
| -- | -- |
|![image](https://media.github.cds.internal.unity3d.com/user/2154/files/687c266a-17f8-488c-892e-34bd0d599021)|![image](https://media.github.cds.internal.unity3d.com/user/2154/files/1504df45-1b30-4515-8f55-6303c79f1e24)|
|![image](https://media.github.cds.internal.unity3d.com/user/2154/files/e17a5606-b162-4800-9b72-5d3258605e89)|![image](https://media.github.cds.internal.unity3d.com/user/2154/files/a728f5ee-b947-49c2-8a2b-9d2c493076d2)|
|![image](https://media.github.cds.internal.unity3d.com/user/2154/files/b88db353-d285-4d01-8d1a-dd279795179e)|![image](https://media.github.cds.internal.unity3d.com/user/2154/files/1a3b2b31-f81e-467b-a6c1-5f15ab41b977)|


Additionally, the atmospheric scattering on opaques is computed differently depending on the option
- When on (default), a 3D LUT is computed at each frame to store the atmospheric attenuation
- When off, it is computed based on the various sky LUTs
The first option is more performant (1ms on the fog pass, and additional savings for transparents), but it only works up to 128km after which the slices of the LUT create visible artifacts

| 3D LUT | 4D sky LUTs |
| -- | -- |
|![image](https://media.github.cds.internal.unity3d.com/user/2154/files/d4aed412-cd80-46d0-bd94-f050064ca465)|![image](https://media.github.cds.internal.unity3d.com/user/2154/files/687c266a-17f8-488c-892e-34bd0d599021)|

I updated the doc accordingly

More comparision shots

| branch | result |
| -- | -- |
|trunk |![image](https://media.github.cds.internal.unity3d.com/user/2154/files/36e4228a-41b8-40ee-a17f-e5cc53e5a195)|
|new default|![image](https://media.github.cds.internal.unity3d.com/user/2154/files/7b59184d-b7c5-4114-b2a8-1dfde0220efa)|
|shader config|![image](https://media.github.cds.internal.unity3d.com/user/2154/files/ba9f2421-8a32-42dc-83b6-815e804b2a84)|