[URP] - Fix Incorrect Viewport Size in CopyDepthPass
This change fixes an issue where the render target viewport size was being changed to an incorrect value during the CopyDepthPass.

In cases where dynamic resolution is active, the camera target descriptor does not reflect the rendered size of the image so it should not be used to control the viewport.

Also, if we're not writing to the backbuffer, then there's no need to change the existing viewport dimensions when performing the blit.

This change resolves the issue by removing the code that updates the viewport size in the case where we aren't writing to the backbuffer.
Note: This is also how RenderingUtils.FinalBlit handles the viewport update logic.