[UUM-45496] Fix debug input errors when emptying input actions
Fix https://jira.unity3d.com/browse/UUM-45496

This PR fixes a bug where errors would be logged in console after domain reload when 
1) user goes to Project Settings > Player and selects Active Input Handling: Input System Package (New), 
2) user goes to Project Settings > Input Manager and deletes all entries in the list, but
3) does not (yet) install the Input System Package through Package Manager.

In this situation (which is inherently incorrect because input is misconfigured), the debug action registration code gave errors after domain reload due to incorrect indexing of the input axes array. Code is fixed to not rely on existing elements in that array.