[VFX] Fix StripData buffer initialization
With the changes for [particle strips missing functionality](https://github.cds.internal.unity3d.com/unity/unity/pull/37417), we created a regression when using strips with instancing.

The issue was not easily noticeable at first, because it involves reusing the same instance:
 - Create some instances of the same VFX
 - Delete some of them
 - Create new ones

When an instance is reused with this issue, you could observe that it may have less particles, and it could even make other instances lose particles.

The cause for this issue was an offset that was not updated to account for the new space added in a buffer, which caused the initialization to access the wrong instances.