[Port] [2023.2] Spotlight culling fix in ortographic camera
The bug is that spotlight culling is incorrect in orthographic cameras, depending on spotlight/camera rotation.

The source of the problem is that circleUp can result in a NaN.
This seems to only happen due to BurstCompile param FloatMode.Fast.
The bug is fixed by using FloatMode.Default instead.

Before this fix, spotlight tiling can look like this:

![image](https://media.github.cds.internal.unity3d.com/user/2726/files/d02ecf66-8840-4853-940a-85f08e38996e)

when it should look like this:

![image](https://media.github.cds.internal.unity3d.com/user/2726/files/5e126559-069a-40b4-b002-369935a73d6e)