[HDRP] - Fix Shadow Flickering With Dynamic Resolution
This change fixes some incorrect logic in the shadow rendering code that
was using the camera's state on the render graph execution timeline to
restore the gpu-side copy of the global constant buffer. The camera's
state changes during the render graph build process, so the state needs
to be captured at build time for correctness.

In this particular case, the camera state doesn't actually change during
shadow rendering so we can also skip the cpu-side global constant buffer
update logic and just push the current cpu-side copy to the gpu.