[APV] Fix scenes always registering in global settings
APV is automatically registering scenes in the global settings when apv is on, even when the scene doesn't have apv
This is a behvaiour we removed in later versions but didn't backport. But it's causing troubles as it creates merge conflicts in the global settings for no reason.
So i changed the code so that it matches what we do on trunk, tested a bit to create new scene, and add them to baking sets and didn't see any issue.