[Backport][2023.2][UUM-53816][SRPCore]Fixing NullReferenceException when trying to display VolumeComponents overrides.
The VolumeManager initializes the volumes for a pipeline when it is initialized.
In the user project, even if the scene view was open, the UniversalRenderPipeline was not created as there was a Built-in debug mode active. When adding an Override on the Volume Profile, the list was null and it has never been initialized.