[RPW] Fixed GlobalSettingsHelpBox for HDRP and updated OpenAndScroll api
https://jira.unity3d.com/browse/UUM-58377

Some of the UI in HDRP have a special FixMe HelpBox when you enable feature on the scene but not in the settings.
It will show a button to help you navigate a settings that needs to be enabled.
After we switched to new Graphics/Global Settings it stopped working. We fix it in this PR.

Here's a gif of HelpBox for VolumetricClouds.
![HelpBoxForVolumetricClouds](https://media.github.cds.internal.unity3d.com/user/5932/files/947253d2-985c-4d9e-932c-eeec1b720c9a)

New OpenAndScrollTo api with IRenderPipelineGraphicsSettings use templates now.