[2022.3] [HDRP] Fix overlay UI issues when rendering to texture being read
On HDRP, without HDR output, the screen space overlay UI is rendered by a pass after `SetFinalTarget` directly on the camera target (as opposed to offscreen when HDR output is enabled). If that camera target is a render texture and the UI is reading from it while it is bound as a render target, there were issues on DX12 (flickering) and PS5 (artefacts).

This didn't happen on 2023.2+ since the overlay UI support was extended and fixed in #29279. This PR grafts only the required change, that says to not render overlay UI on offscreen camera. This has been the behaviour on BiRP (since the engine rendered the UI on the backbuffer directly only at the end of frame) and is the behaviour in URP.