Test fix for BatchRendererGroup_HDRP
Remove the physically based sky component from the BRGCullingFlags test scene, in order to match the existing test image.

This started failing after the merge of #43236, but is due to a bugfix related to the ambient mode when the physical sky is active.  Removing the physical sky sets the ambient lighting back to black as expected by the existing test images.