Allow to run test on BRG - ErrorMaterial scene
The goal of this PR is to allow for graphics tests to run on the ErrorMaterial scene, despite shader errors that causes job failures in Yamato. The solution has 2 parts:

1 - Ignore failing messages using UTF LogAssert API (see GraphicsTests.cs)

2 - Disable scene in build profile when running the build job in Yamato (see hdrp-brg.metafile)

UTR does not allow to ignore failing messages (only compilation errors), this is why I opted to disable the scene in the BuildSettings.