Improve HDRP color pyramid speed
Add a code path to the compute shader for the color pyramid on compute prefered devices (k_PreferFragment = false)
On Switch, the pixel path is still a lot faster (2.5ms VS 4.5ms) so I left the pixel path for Switch.

Before:
![image](https://media.github.cds.internal.unity3d.com/user/745/files/41514b43-c430-4da3-ba8a-76186c462740)

After:
![image](https://media.github.cds.internal.unity3d.com/user/745/files/539bc0af-a9ca-4aed-87da-12cb9e7b4932)

Note that we still use a regular Blit for the copy of the first mip because on certain format like R11G11B10, the pixel shader is faster than a compute to output the texture data (fragments have HW export that can do the convertion).

I also removed one temporary render target that was used for the multi-pass blur and not pooled by RenderGraph.

I experimented merging of the compute passes like doing several mip generation in a single dispatch using LDS but it didn't make a big difference and the code was a lot less readable so I decided it was not worth it. Similar story with the merged gaussian blur pass (1 dispatch for all the mips) + it'd have decrease the quality of the downsample.